feat(web): consolidate revalidation into single endpoint
New /api/revalidate route handles both ?tag= and ?path= params.
Old /api/revalidate/featured-artists becomes a 308 redirect to preserve
backward compatibility during deployment.
